requirements.txt
Flask
Flask-SQLAlchemy
jinja2
PyYAML
requests
python-dotenv
pokebase